inplace_function now builds, basic tests added, cmake file updated
