+++ SG14_test/SG14_test.h
@@ -5,10 +5,10 @@ namespace sg14_test
 {
 	void static_ring_test();
 	void dynamic_ring_test();
-	void thread_communication_test();
 	void unstable_remove_test();
 	void uninitialized();
-	void plf_test_suite();
+  void plf_test_suite();
+  void inplace_function_test();
 }
 
 #endif
